+++ src/main/java/com/zematix/jworldcup/backend/mapper/MatchMapper.java
@@ -7,13 +7,4 @@ import com.zematix.jworldcup.backend.entity.Match;
 
 @Mapper(uses=RoundMapper.class)
 public interface MatchMapper extends MapperBase<MatchDto, Match> {
-
-//	@Override
-//	@Mapping(target = "round", source = "round", mappingControl =  )
-//	MatchDto entityToDto(Match match);
-//
-//	@Override
-//	@Mapping(target = "round", source = "round") 
-//	Match dtoToEntity(MatchDto matchDto);
-
 }