+++ src/main/java/com/zematix/jworldcup/backend/service/BetService.java
@@ -74,6 +74,22 @@ public class BetService extends ServiceBase {
 			bet.setScore(retrieveScoreByBet(bet, favouriteTeamIds));
 			bet.setFavouriteTeamIndex(retrieveFavouriteTeamIndexByBet(bet, favouriteTeamIds));
 		}
+		
+		// load lazy associations
+		for (Bet bet: bets) {
+			Match match = bet.getMatch();
+			if (match.getTeam1() != null) {
+				match.getTeam1().getName();
+				match.getTeam1().getGroup().getName();
+			}
+			if (match.getTeam2() != null) {
+				match.getTeam2().getName();
+				match.getTeam2().getGroup().getName();
+			}
+			bet.getUser().getLoginName();
+			bet.getUser().getRoles().size();
+		}
+
 		return bets;
 	}
 	
@@ -93,6 +109,19 @@ public class BetService extends ServiceBase {
 			errMsgs.add(ParameterizedMessage.create("MISSING_TIP"));
 			throw new ServiceException(errMsgs);
 		}
+
+		// load lazy associations
+		Match match = bet.getMatch();
+		if (match.getTeam1() != null) {
+			match.getTeam1().getName();
+			match.getTeam1().getGroup().getName();
+		}
+		if (match.getTeam2() != null) {
+			match.getTeam2().getName();
+			match.getTeam2().getGroup().getName();
+		}
+		bet.getUser().getLoginName();
+		bet.getUser().getRoles().size();
 
 		bet.setScore(retrieveScoreByBet(bet, null));
 		bet.setFavouriteTeamIndex(retrieveFavouriteTeamIndexByBet(bet, null));
@@ -114,10 +143,27 @@ public class BetService extends ServiceBase {
 		checkNotNull(userGroupId);
 		
 		List<Bet> bets = betDao.retrieveBetsByMatchAndUserGroup(matchId, userGroupId);
+		
+		// load lazy associations
+		for (Bet bet: bets) {
+			Match match = bet.getMatch();
+			if (match.getTeam1() != null) {
+				match.getTeam1().getName();
+				match.getTeam1().getGroup().getName();
+			}
+			if (match.getTeam2() != null) {
+				match.getTeam2().getName();
+				match.getTeam2().getGroup().getName();
+			}
+			bet.getUser().getLoginName();
+			bet.getUser().getRoles().size();
+		}
+		
 		for (Bet bet: bets) {
 			bet.setScore(retrieveScoreByBet(bet, null));
 			bet.setFavouriteTeamIndex(retrieveFavouriteTeamIndexByBet(bet, null));
 		}
+		
 		return bets;
 	}
 	
@@ -186,6 +232,12 @@ public class BetService extends ServiceBase {
 			}
 		}
 
+		if (bet != null) {
+			// load lazy associations
+			bet.getUser().getLoginName();
+			bet.getUser().getRoles().size();
+		}
+		
 		if (bet != null) {
 			bet.setScore(this.retrieveScoreByBet(bet, null));
 			bet.setFavouriteTeamIndex(retrieveFavouriteTeamIndexByBet(bet, null));
@@ -295,14 +347,8 @@ public class BetService extends ServiceBase {
 		checkNotNull(bet);
 		Integer favouriteTeamIndex = null;
 		
-		Long favouriteGroupTeamId = null;
-		Long favouriteKnockoutTeamId = null;
 		if (favouriteTeamIds == null) {
 			favouriteTeamIds = retrieveFavouriteTeamIdsByBet(bet);
-		}
-		if (favouriteTeamIds != null) { 
-			favouriteGroupTeamId = favouriteTeamIds.getValue1();
-			favouriteKnockoutTeamId = favouriteTeamIds.getValue2();
 		}
 
 		boolean isGroupmatch = Boolean.TRUE.equals(bet.getMatch().getRound().getIsGroupmatchAsBoolean());
@@ -329,16 +375,8 @@ public class BetService extends ServiceBase {
 		checkNotNull(eventId);
 		checkNotNull(userId);
 
-		int score = 0;
-		List<Bet> bets = betDao.retrieveBetsByEventAndUser(eventId, userId);
-		Pair<Long> favouriteTeamIds = null;
-		for (Bet bet : bets) {
-			if (favouriteTeamIds == null) {
-				favouriteTeamIds = retrieveFavouriteTeamIdsByBet(bet);
-			}
-			score += retrieveScoreByBet(bet, favouriteTeamIds);
-		}
-		return score;
+		List<Bet> bets = retrieveBetsByEventAndUser(eventId, userId);
+		return bets.stream().mapToInt(Bet::getScore).sum();
 	}
 
 	/**
@@ -356,14 +394,10 @@ public class BetService extends ServiceBase {
 
 		Map<LocalDateTime, Integer> mapScoreByDate = new HashMap<>();
 		int score = 0;
-		List<Bet> bets = betDao.retrieveBetsByEventAndUser(eventId, userId);
-		Pair<Long> favouriteTeamIds = null;
+		List<Bet> bets = retrieveBetsByEventAndUser(eventId, userId);
 		for (Bet bet : bets) {
-			if (favouriteTeamIds == null) {
-				favouriteTeamIds = retrieveFavouriteTeamIdsByBet(bet);
-			}
 			LocalDateTime matchDate = CommonUtil.truncateDateTime(bet.getMatch().getStartTime());
-			score += retrieveScoreByBet(bet, favouriteTeamIds);
+			score += bet.getScore();
 			mapScoreByDate.put(matchDate, score);
 		}
 		return mapScoreByDate;