+++ src/main/java/com/zematix/jworldcup/backend/service/MatchService.java
@@ -18,22 +18,15 @@ import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.interceptor.SimpleKeyGenerator;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.event.EventListener;
-import org.springframework.lang.NonNull;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
 import com.zematix.jworldcup.backend.configuration.CachingConfig;
 import com.zematix.jworldcup.backend.dao.CommonDao;
 import com.zematix.jworldcup.backend.dao.MatchDao;
-import com.zematix.jworldcup.backend.dao.RoundDao;
 import com.zematix.jworldcup.backend.emun.ParameterizedMessageType;
 import com.zematix.jworldcup.backend.entity.Event;
 import com.zematix.jworldcup.backend.entity.Group;
@@ -55,7 +48,6 @@ import com.zematix.jworldcup.backend.util.LambdaExceptionUtil;
  * However it may also inject other services and DAO classes.
  */
 @Service
-@Configuration
 @Transactional
 public class MatchService extends ServiceBase {
 
@@ -66,7 +58,7 @@ public class MatchService extends ServiceBase {
 	private MatchDao matchDao;
 
 	@Inject 
-	private RoundDao roundDao;
+	private RoundService roundService;
 
 	@Inject 
 	private CommonDao commonDao;
@@ -113,16 +105,16 @@ public class MatchService extends ServiceBase {
 
 		checkNotNull(eventId);
 		
-		List<Round> rounds = roundDao.retrieveRoundsByEvent(eventId);
+		List<Round> rounds = roundService.retrieveRoundsByEvent(eventId);
 		
 		// load lazy entity associations
 		for (Round round : rounds) {
 			for (Match match : round.getMatches()) {
 				if (match.getTeam1() != null) {
-					match.getTeam1().getTeamId();
+					match.getTeam1().getName();
 				}
 				if (match.getTeam2() != null) {
-					match.getTeam2().getTeamId();
+					match.getTeam2().getName();
 				}
 			}
 		}
@@ -153,12 +145,14 @@ public class MatchService extends ServiceBase {
 		
 		// load lazy associations
 		for (Match match : matches) {
-			match.getRound().getRoundId();
+			match.getRound().getName();
 			if (match.getTeam1() != null) {
-				match.getTeam1().getTeamId();
+				match.getTeam1().getName();
+				match.getTeam1().getGroup().getName();
 			}
 			if (match.getTeam2() != null) {
-				match.getTeam2().getTeamId();
+				match.getTeam2().getName();
+				match.getTeam2().getGroup().getName();
 			}
 			if (match.getTeam1() != null && match.getTeam2() != null) {
 				match.setResultSignByTeam1(getMatchResult(match, match.getTeam1().getTeamId()));
@@ -173,6 +167,27 @@ public class MatchService extends ServiceBase {
 	}
 
 	/**
+	 * Returns a list of {@link Match} instances where each match element has the given 
+	 * {@code teamId} participant, the match is finished (has valid result) and 
+	 * the match is played in the group stage.
+	 *
+	 * @param teamId
+	 * @return list of {@link Match} instances containing all finished matches belongs to the given team
+	 * @throws IllegalArgumentException if any of the parameters is null
+	 */
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public List<Match> retrieveFinishedGroupMatchesByTeam(Long teamId) {
+		List<Match> matches = matchDao.retrieveFinishedGroupMatchesByTeam(teamId);
+		
+		// load lazy associations
+		for (Match match : matches) {
+			match.getRound().getName();
+		}
+		
+		return matches;
+	}
+	
+	/**
 	 * Retrieves a list of {@link Match} instances belongs to the  given {@code eventId}, 
 	 * with not null {@link Match#participantsRule} value, located in the 
 	 * knockout stage and has at least one {@link Team} participant with {@code null} value.
@@ -185,9 +200,6 @@ public class MatchService extends ServiceBase {
 	 */
 	@Transactional(readOnly = true)
 	public List<Match> retrieveMatchesWithoutParticipantsByEvent(Long eventId) throws ServiceException {
-
-		List<ParameterizedMessage> errMsgs = new ArrayList<>();
-
 		checkNotNull(eventId);
 		
 		List<Match> matches = matchDao.retrieveMatchesWithoutParticipantsByEvent(eventId);
@@ -195,18 +207,29 @@ public class MatchService extends ServiceBase {
 		// load lazy associations
 		for (Match match : matches) {
 			if (match.getTeam1() != null) {
-				match.getTeam1().getTeamId();
+				match.getTeam1().getName();
 			}
 			if (match.getTeam2() != null) {
-				match.getTeam2().getTeamId();
+				match.getTeam2().getName();
 			}
 		}
 		
-		if (!errMsgs.isEmpty()) {
-			throw new ServiceException(errMsgs);
-		}
-
 		return matches;
+	}
+
+	/**
+	 * Returns a list of {@link Match#participantsRule} values of all matches
+	 * belongs to the provided {@code eventId} event. Only the non empty
+	 * values are retrieved of the knock-out matches where there is at least
+	 * a missing team participant.
+	 * 
+	 * @param eventId
+	 * @return list of participant rules of all matches of the provided event
+	 * @throws IllegalArgumentException if any of the parameters is null
+	 */
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public List<String> retrieveParticipantRulesOfMatchesByEvent(Long eventId) {
+		return matchDao.retrieveParticipantRulesOfMatchesByEvent(eventId);
 	}
 
 	/**
@@ -230,16 +253,16 @@ public class MatchService extends ServiceBase {
 		}
 		
 		// load lazy associations
-		match.getRound().getRoundId();
+		match.getRound().getName();
 		if (match.getTeam1() != null) {
-			match.getTeam1().getTeamId();
+			match.getTeam1().getName();
+			match.getTeam1().getGroup().getName();
 		}
 		if (match.getTeam2() != null) {
-			match.getTeam2().getTeamId();
+			match.getTeam2().getName();
+			match.getTeam2().getGroup().getName();
 		}
 
-		commonDao.detachEntity(match);
-		
 		return match;
 	}
 
@@ -333,7 +356,7 @@ public class MatchService extends ServiceBase {
 		}
 
 		// retrieve match from table and validate it
-		match = commonDao.findEntityById(Match.class, matchId);
+		match = retrieveMatch(matchId);
 		if (match == null) {
 			errMsgs.add(ParameterizedMessage.create("MISSING_MATCH"));
 			throw new ServiceException(errMsgs);
@@ -357,43 +380,18 @@ public class MatchService extends ServiceBase {
 		
 		// After successful transaction commit after the end of this method a new transaction
 		// is called asynchronously by event handler, which might update the upcoming matches. 
-		// A wrapper service method calling this method and the mentioned asynchronous method 
-		// might be easier, but it needs an extra service. 
-		final Match finalMatch = match;
-		executeAfterTransactionCommits(() -> {
-			// launches event to generate a header message about the updated match elsewhere
-			PublishedEvent<Match> event = new PublishedEvent<>(finalMatch, true);
-			applicationEventPublisher.publishEvent(event);			
-		});
+		applicationEventPublisher.publishEvent(new PublishedEvent<>(match));
 		
 		return match;
 	}
 	
-	private void executeAfterTransactionCommits(Runnable task) {
-		TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-			@Override
-			public void afterCommit() {
-				task.run();
-			}
-		});
-	}
-	
 	/**
-	 * Invoked from {@link MatchService#saveMatch(Long, boolean, Boolean, LocalDateTime, Byte, Byte, Byte, Byte, Byte, Byte)
+	 * Invoked from {@link MatchAsyncService#onUpdateMatchEvent(PublishedEvent<Match>)
 	 * when a match result is saved.
-	 * @param event - contains the saved match
+	 * @param match - saved match
 	 */
-	@Transactional(propagation = Propagation.REQUIRES_NEW)
-	@Async
-	@EventListener(condition = "#event.success")
-	public void onUpdateMatchEvent(@NonNull PublishedEvent/*<Match>*/ event) throws ServiceException {
-		if (!event.getEntity().getClass().equals(Match.class)) {
-			return;
-		}
-		
-		Match match = (Match)event.getEntity();
-		logger.info("onUpdateMatchEvent matchId: {}", match.getMatchId());
-		
+	@Transactional(propagation = Propagation.MANDATORY)
+	public void onUpdateMatchEvent(Match match) throws ServiceException {
 		updateMatchParticipants(match.getEvent().getEventId(), match.getMatchId());
 		// update cached value
 		applicationService.refreshEventCompletionPercentCache(match.getEvent().getEventId());
@@ -434,7 +432,7 @@ public class MatchService extends ServiceBase {
 		}
 
 		// retrieve match from table and validate it
-		match = commonDao.findEntityById(Match.class, matchId);
+		match = retrieveMatch(matchId);
 		if (match == null) {
 			errMsgs.add(ParameterizedMessage.create("MISSING_MATCH"));
 			throw new ServiceException(errMsgs);
@@ -456,14 +454,7 @@ public class MatchService extends ServiceBase {
 		
 		// After successful transaction commit after the end of this method a new transaction
 		// is called asynchronously by event handler, which might update the upcoming matches. 
-		// A wrapper service method calling this method and the mentioned asynchronous method 
-		// might be easier, but it needs an extra service. 
-		final Match finalMatch = match;
-		executeAfterTransactionCommits(() -> {
-			// launches event to generate a header message about the updated match elsewhere
-			PublishedEvent<Match> event = new PublishedEvent<>(finalMatch, true);
-			applicationEventPublisher.publishEvent(event);			
-		});
+		applicationEventPublisher.publishEvent(new PublishedEvent<>(match));
 		
 		return match;
 	}
@@ -1145,12 +1136,12 @@ public class MatchService extends ServiceBase {
 		
 		// load lazy associations
 		for (Match match : matches) {
-			match.getRound().getRoundId();
+			match.getRound().getName();
 			if (match.getTeam1() != null) {
-				match.getTeam1().getTeamId();
+				match.getTeam1().getName();
 			}
 			if (match.getTeam2() != null) {
-				match.getTeam2().getTeamId();
+				match.getTeam2().getName();
 			}
 		}
 		
@@ -1233,7 +1224,7 @@ public class MatchService extends ServiceBase {
 		
 		// update match table
 
-		match = commonDao.findEntityById(Match.class, matchId);
+		match = retrieveMatch(matchId);
 		if (match == null) {
 			errMsgs.add(ParameterizedMessage.create("MISSING_MATCH"));
 			throw new ServiceException(errMsgs);