+++ src/main/java/com/zematix/jworldcup/backend/service/UserService.java
@@ -314,6 +314,8 @@ public class UserService extends ServiceBase {
 
 		try {
 			user = userDao.modifyUser(user, fullName, newEmailAddr, encryptedNewLoginPassword, zoneId, applicationService.getActualDateTime());
+			
+			user.getRoles().size(); // lazy fetch
 		}
 		catch (Exception e) {
 			errMsgs.add(ParameterizedMessage.create("DB_SAVE_FAILED"));
@@ -420,6 +422,7 @@ public class UserService extends ServiceBase {
 		User user = userDao.findUserByLoginName(loginName);
 		if (user != null) {
 			user.getRoles().size(); // forced lazy fetch
+			user.getUserStatus().getUserStatusUsers().size();
 		}
 		
 		return user;
@@ -478,6 +481,7 @@ public class UserService extends ServiceBase {
 			throw new ServiceException(errMsgs);
 		}
 
+		user.getRoles().size(); // lazy fetch
 		commonDao.detachEntity(user);
 		
 		return user;
@@ -490,6 +494,7 @@ public class UserService extends ServiceBase {
 	 * 
 	 * @return a map containing all supported time zone key/value pairs
 	 */
+	@Cacheable(cacheNames = CachingConfig.CACHE_TIMEZONE_IDS)
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public Map<String, String> getAllSupportedTimeZoneIds() {
 		List<String> zoneList = new ArrayList<>(ZoneId.getAvailableZoneIds());