+++ src/test/resources/database/controller/jwt-authentication-controller-after.sql
@@ -1,0 +1,3 @@
+-- saveUser
+DELETE FROM user__role WHERE role_id=3 AND user_id=6;
+DELETE FROM user WHERE user_id = 6;