+++ CHANGES.md
@@ -1,5 +1,10 @@
 ## Changes
 
+### Version 1.1.1 / 2024-02-??
+
+* Feature: cached retrieval of static timezone ids
+* Fix: open-in-view got false setting => lazy fetch
+
 ### Version 1.1.0 / 2024-01-13
 
 * Fix: undetected event (tournament) change if the latest one was involved